Converts prioritization solution into a more digestible output

This function converts the prioritization solution that uses patches to a raster or sf object (depending on the format of your spatial_grid) that clearly marks areas suggested for protection (1) and areas not suggested for protection (0)

Usage

convert_solution(solution, patch_df, spatial_grid)

Arguments

solution

the solution that results from using prioritizr::solve() on prioritizr::problem()

patch_df

a dataframe generated by create_patch_df() that includes constraints for each patch and grid cell combination

spatial_grid

a raster or sf template with the desired resolution and coordinate reference system generated by spatialgridr::get_grid(); values in areas of interest are 1, while all other values are NA (only required if feature is a sf object)

Value

A raster or sf object that denotes areas suggested for protection (value = 1)

Examples

# Start with a little housekeeping to get the data from oceandatr
# Choose area of interest (Bermuda EEZ)
area <- oceandatr::get_area(area_name = "Bermuda",  mregions_column = "territory1")
projection <-'+proj=laea +lon_0=-64.8108333 +lat_0=32.3571917 +datum=WGS84 +units=m +no_defs'
# Create a planning grid
planning_raster <- spatialgridr::get_grid(area, projection = projection)
# Grab all relevant data
features_raster <- oceandatr::get_features(spatial_grid = planning_raster)
#> Getting depth zones...
#> This may take seconds to minutes, depending on grid size
#> Getting seamount data...
#> Spherical geometry (s2) switched off
#> although coordinates are longitude/latitude, st_intersection assumes that they
#> are planar
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
#> Spherical geometry (s2) switched on
#> Getting knoll data...
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
#> Getting geomorphology data...
#> Getting coral data...
#> |-- Coral data found for antipatharia coral, cold coral, octocoral
#> Getting environmental regions data... This could take several minutes
# Separate seamount data - we want to protect entire patches
seamounts_raster <- features_raster[["seamounts"]]
features_raster <- features_raster[[names(features_raster)[names(features_raster) != "seamounts"]]]
# Create a "cost" to protecting a cell - just a uniform cost for this example
cost_raster <- stats::setNames(planning_raster, "cost")
# Create patches from layer
patches_raster <- create_patches(seamounts_raster)
# Create patch dataframe
patches_raster_df <- create_patch_df(spatial_grid = planning_raster, features = features_raster,
  patches = patches_raster, costs = cost_raster)
#> [1] "Processing patch 1 of 7"
#> [1] "Processing patch 2 of 7"
#> [1] "Processing patch 3 of 7"
#> [1] "Processing patch 4 of 7"
#> [1] "Processing patch 5 of 7"
#> [1] "Processing patch 6 of 7"
#> [1] "Processing patch 7 of 7"
# Create boundary matrix for prioritizr
boundary_matrix <- create_boundary_matrix(spatial_grid = planning_raster, patches = patches_raster,
  patch_df = patches_raster_df)
# Create target features - using just 20% for every feature
features_targets <- features_targets(targets = rep(0.2, (terra::nlyr(features_raster)) + 1),
  features = features_raster, pre_patches = seamounts_raster)
# Create constraint targets
constraint_targets <- constraints_targets(feature_targets = features_targets,
  patch_df = patches_raster_df)
# Create prioritization problem
problem_raster <- prioritizr::problem(x = patches_raster_df,
  features = constraint_targets$feature, cost_column = "cost") %>%
  prioritizr::add_min_set_objective() %>%
  prioritizr::add_manual_targets(constraint_targets) %>%
  prioritizr::add_binary_decisions() %>%
  prioritizr::add_boundary_penalties(penalty = 0.000002, data = boundary_matrix) %>%
  prioritizr::add_gurobi_solver(gap = 0.1, threads = parallel::detectCores()-1)
#> Warning: `data` has unexpected values.
#> ✖ If `data` is from an older version of prioritizr, then use
#>   `boundary_matrix()` to recreate it.
#> ℹ `x` might have spatially overlapping planning units.
# Solve problem
solution <- solve(problem_raster)
#> Gurobi Optimizer version 10.0.0 build v10.0.0rc2 (mac64[arm])
#> 
#> CPU model: Apple M1 Max
#> Thread count: 10 physical cores, 10 logical processors, using up to 9 threads
#> 
#> Optimize a model with 76295 rows, 56073 columns and 213985 nonzeros
#> Model fingerprint: 0x3f39d706
#> Variable types: 0 continuous, 56073 integer (56073 binary)
#> Coefficient statistics:
#>   Matrix range     [1e+00, 5e+02]
#>   Objective range  [2e-02, 5e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [1e+00, 4e+03]
#> Found heuristic solution: objective 3733.6350000
#> Presolve removed 952 rows and 317 columns
#> Presolve time: 0.48s
#> Presolved: 75343 rows, 55756 columns, 212535 nonzeros
#> Variable types: 0 continuous, 55756 integer (55756 binary)
#> Root relaxation presolved: 75343 rows, 55756 columns, 212535 nonzeros
#> 
#> Deterministic concurrent LP optimizer: primal and dual simplex
#> Showing first log only...
#> 
#> Warning: Markowitz tolerance tightened to 0.5
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>    84397    3.7261655e+03   0.000000e+00   4.375845e+04      5s
#>    95116    3.7247900e+03   0.000000e+00   3.433822e+03     10s
#>   102028    3.7219391e+03   0.000000e+00   1.232931e+03     15s
#>   108598    3.7216986e+03   0.000000e+00   2.682192e+03     20s
#>   115266    3.7214453e+03   0.000000e+00   2.536798e+04     25s
#>   120908    3.7212532e+03   0.000000e+00   1.725392e+05     30s
#>   127322    3.7211680e+03   0.000000e+00   1.195761e+04     35s
#>   130963    3.7211541e+03   0.000000e+00   6.248709e+04     40s
#>   134868    3.7211099e+03   0.000000e+00   8.665098e+04     45s
#>   138516    3.7210725e+03   0.000000e+00   6.396256e+02     50s
#> Concurrent spin time: 0.00s
#> 
#> Solved with dual simplex
#> 
#> Root simplex log...
#> 
#> Iteration    Objective       Primal Inf.    Dual Inf.      Time
#>    73586    3.7211523e+03   0.000000e+00   0.000000e+00     52s
#> 
#> Root relaxation: objective 3.721152e+03, 73586 iterations, 51.10 seconds (199.15 work units)
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 3721.15230    0 53807 3733.63500 3721.15230  0.33%     -   51s
#> 
#> Explored 1 nodes (73708 simplex iterations) in 51.93 seconds (200.83 work units)
#> Thread count was 9 (of 10 available processors)
#> 
#> Solution count 1: 3733.64 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 3.733635000000e+03, best bound 3.721152297269e+03, gap 0.3343%
# Convert to a more digestible format
suggested_protection <- convert_solution(solution = solution, patch_df = patches_raster_df,
  spatial_grid = planning_raster)